# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DevPulse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  githubId: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface GitHubRepo {
  id: number;
  name: string;
  fullName: string;
  /** @nullable */
  description?: string | null;
  private: boolean;
  /** @nullable */
  language?: string | null;
  updatedAt?: string;
}

export interface GitHubCommit {
  sha: string;
  message: string;
  author: string;
  date: string;
}

export interface GitHubPR {
  id: number;
  number: number;
  title: string;
  state: string;
  createdAt: string;
  updatedAt: string;
}

export interface GitHubIssue {
  id: number;
  number: number;
  title: string;
  state: string;
  createdAt: string;
}

export interface GitHubActivity {
  commits: GitHubCommit[];
  prs: GitHubPR[];
  issues: GitHubIssue[];
}

export interface VelocityDay {
  date: string;
  commits: number;
  prs: number;
}

export interface Standup {
  id: number;
  userId: number;
  repoName: string;
  date: string;
  yesterday: string;
  today: string;
  blockers: string[];
  /** @nullable */
  nextPriorityTask?: string | null;
  /** @nullable */
  velocityScore?: number | null;
  sentToSlack: boolean;
  createdAt: string;
}

export type PublicStandupAuthor = {
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
};

export interface PublicStandup {
  id: number;
  repoName: string;
  date: string;
  yesterday: string;
  today: string;
  blockers: string[];
  /** @nullable */
  nextPriorityTask?: string | null;
  /** @nullable */
  velocityScore?: number | null;
  createdAt: string;
  author: PublicStandupAuthor;
}

export interface GenerateStandupBody {
  repoName: string;
  /** ISO timestamp (defaults to 24h ago) */
  since?: string;
}

export type PreferencesStandupTone =
  (typeof PreferencesStandupTone)[keyof typeof PreferencesStandupTone];

export const PreferencesStandupTone = {
  professional: "professional",
  casual: "casual",
  "bullet-points": "bullet-points",
} as const;

export interface Preferences {
  userId: number;
  /** @nullable */
  slackWebhookUrl?: string | null;
  standupTone: PreferencesStandupTone;
  /** @nullable */
  defaultRepo?: string | null;
}

export type UpdatePreferencesBodyStandupTone =
  (typeof UpdatePreferencesBodyStandupTone)[keyof typeof UpdatePreferencesBodyStandupTone];

export const UpdatePreferencesBodyStandupTone = {
  professional: "professional",
  casual: "casual",
  "bullet-points": "bullet-points",
} as const;

export interface UpdatePreferencesBody {
  /** @nullable */
  slackWebhookUrl?: string | null;
  standupTone?: UpdatePreferencesBodyStandupTone;
  /** @nullable */
  defaultRepo?: string | null;
}

export interface TestSlackBody {
  webhookUrl: string;
}

export type CoachChatMessageRole =
  (typeof CoachChatMessageRole)[keyof typeof CoachChatMessageRole];

export const CoachChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface CoachChatMessage {
  role: CoachChatMessageRole;
  content: string;
}

export interface CoachChatBody {
  message: string;
  conversationHistory?: CoachChatMessage[];
  /** Optional repo to fetch activity context for */
  repo?: string;
}

export interface CoachChatResponse {
  reply: string;
}

export interface LinkedinPostBody {
  standupId: number;
}

export interface LinkedinPostResponse {
  post: string;
}

export interface StreakData {
  currentStreak: number;
  longestStreak: number;
}

export interface DayVelocity {
  day: string;
  avgVelocity: number;
  count: number;
}

export interface PrCycleTime {
  avgHours: number;
  label: string;
  color: string;
}

export interface Keyword {
  word: string;
  count: number;
}

export type GetGithubActivityParams = {
  /**
   * Full repo name (owner/repo)
   */
  repo: string;
  /**
   * ISO timestamp to fetch activity since
   */
  since?: string;
};

export type GetGithubVelocityParams = {
  /**
   * Full repo name (owner/repo)
   */
  repo: string;
};

export type GetInsightsPrCycleTimeParams = {
  /**
   * Full repo name (owner/repo)
   */
  repo?: string;
};