# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DevPulse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coach/chat:
    post:
      operationId: coachChat
      tags: [coach]
      summary: Chat with AI Sprint Coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachChatBody"
      responses:
        "200":
          description: Coach reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachChatResponse"
        "401":
          description: Not authenticated
  /share/linkedin-post:
    post:
      operationId: generateLinkedinPost
      tags: [share]
      summary: Generate a LinkedIn post from a standup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkedinPostBody"
      responses:
        "200":
          description: Generated LinkedIn post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkedinPostResponse"
        "401":
          description: Not authenticated
  /insights/streak:
    get:
      operationId: getInsightsStreak
      tags: [insights]
      summary: Get current commit streak
      responses:
        "200":
          description: Streak data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreakData"
        "401":
          description: Not authenticated
  /insights/best-day:
    get:
      operationId: getInsightsBestDay
      tags: [insights]
      summary: Get best day of week by velocity
      responses:
        "200":
          description: Best day analysis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayVelocity"
        "401":
          description: Not authenticated
  /insights/pr-cycle-time:
    get:
      operationId: getInsightsPrCycleTime
      tags: [insights]
      summary: Get average PR cycle time
      parameters:
        - name: repo
          in: query
          required: false
          schema:
            type: string
          description: Full repo name (owner/repo)
      responses:
        "200":
          description: PR cycle time data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrCycleTime"
        "401":
          description: Not authenticated
  /insights/top-keywords:
    get:
      operationId: getInsightsTopKeywords
      tags: [insights]
      summary: Get top commit message keywords
      responses:
        "200":
          description: Top keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
        "401":
          description: Not authenticated
components:
  schemas:
...
    CoachChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content
    CoachChatBody:
      type: object
      properties:
        message:
          type: string
        conversationHistory:
          type: array
          items:
            $ref: "#/components/schemas/CoachChatMessage"
        repo:
          type: string
          description: Optional repo to fetch activity context for
      required:
        - message
    CoachChatResponse:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
    LinkedinPostBody:
      type: object
      properties:
        standupId:
          type: integer
      required:
        - standupId
    LinkedinPostResponse:
      type: object
      properties:
        post:
          type: string
      required:
        - post
    StreakData:
      type: object
      properties:
        currentStreak:
          type: integer
        longestStreak:
          type: integer
      required:
        - currentStreak
        - longestStreak
    DayVelocity:
      type: object
      properties:
        day:
          type: string
        avgVelocity:
          type: number
        count:
          type: integer
      required:
        - day
        - avgVelocity
        - count
    PrCycleTime:
      type: object
      properties:
        avgHours:
          type: number
        label:
          type: string
        color:
          type: string
      required:
        - avgHours
        - label
        - color
    Keyword:
      type: object
      properties:
        word:
          type: string
        count:
          type: integer
      required:
        - word
        - count