# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DevPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetAuthMeResponse = zod.object({
  id: zod.number(),
  githubId: zod.string(),
  username: zod.string(),
  avatarUrl: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Logout current user
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List user's GitHub repositories
 */
export const GetGithubReposResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  fullName: zod.string(),
  description: zod.string().nullish(),
  private: zod.boolean(),
  language: zod.string().nullish(),
  updatedAt: zod.string().optional(),
});
export const GetGithubReposResponse = zod.array(GetGithubReposResponseItem);

/**
 * @summary Get GitHub activity for a repository
 */
export const GetGithubActivityQueryParams = zod.object({
  repo: zod.coerce.string().describe("Full repo name (owner\/repo)"),
  since: zod.coerce
    .string()
    .optional()
    .describe("ISO timestamp to fetch activity since"),
});

export const GetGithubActivityResponse = zod.object({
  commits: zod.array(
    zod.object({
      sha: zod.string(),
      message: zod.string(),
      author: zod.string(),
      date: zod.string(),
    }),
  ),
  prs: zod.array(
    zod.object({
      id: zod.number(),
      number: zod.number(),
      title: zod.string(),
      state: zod.string(),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    }),
  ),
  issues: zod.array(
    zod.object({
      id: zod.number(),
      number: zod.number(),
      title: zod.string(),
      state: zod.string(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Get commit velocity for last 7 days
 */
export const GetGithubVelocityQueryParams = zod.object({
  repo: zod.coerce.string().describe("Full repo name (owner\/repo)"),
});

export const GetGithubVelocityResponseItem = zod.object({
  date: zod.string(),
  commits: zod.number(),
  prs: zod.number(),
});
export const GetGithubVelocityResponse = zod.array(
  GetGithubVelocityResponseItem,
);

/**
 * @summary Generate AI standup from GitHub activity
 */
export const GenerateStandupBody = zod.object({
  repoName: zod.string(),
  since: zod
    .string()
    .optional()
    .describe("ISO timestamp (defaults to 24h ago)"),
});

export const GenerateStandupResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  repoName: zod.string(),
  date: zod.string(),
  yesterday: zod.string(),
  today: zod.string(),
  blockers: zod.array(zod.string()),
  nextPriorityTask: zod.string().nullish(),
  velocityScore: zod.number().nullish(),
  sentToSlack: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Get standup history for current user
 */
export const GetStandupHistoryResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  repoName: zod.string(),
  date: zod.string(),
  yesterday: zod.string(),
  today: zod.string(),
  blockers: zod.array(zod.string()),
  nextPriorityTask: zod.string().nullish(),
  velocityScore: zod.number().nullish(),
  sentToSlack: zod.boolean(),
  createdAt: zod.string(),
});
export const GetStandupHistoryResponse = zod.array(
  GetStandupHistoryResponseItem,
);

/**
 * @summary Get a specific standup by ID
 */
export const GetStandupParams = zod.object({
  id: zod.coerce.number(),
});

export const GetStandupResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  repoName: zod.string(),
  date: zod.string(),
  yesterday: zod.string(),
  today: zod.string(),
  blockers: zod.array(zod.string()),
  nextPriorityTask: zod.string().nullish(),
  velocityScore: zod.number().nullish(),
  sentToSlack: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Send a standup to Slack
 */
export const SendStandupToSlackParams = zod.object({
  id: zod.coerce.number(),
});

export const SendStandupToSlackResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get a standup publicly (no auth required)
 */
export const GetPublicStandupParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPublicStandupResponse = zod.object({
  id: zod.number(),
  repoName: zod.string(),
  date: zod.string(),
  yesterday: zod.string(),
  today: zod.string(),
  blockers: zod.array(zod.string()),
  nextPriorityTask: zod.string().nullish(),
  velocityScore: zod.number().nullish(),
  createdAt: zod.string(),
  author: zod.object({
    username: zod.string(),
    avatarUrl: zod.string().nullish(),
  }),
});

/**
 * @summary Get current user preferences
 */
export const GetPreferencesResponse = zod.object({
  userId: zod.number(),
  slackWebhookUrl: zod.string().nullish(),
  standupTone: zod.enum(["professional", "casual", "bullet-points"]),
  defaultRepo: zod.string().nullish(),
});

/**
 * @summary Update user preferences
 */
export const UpdatePreferencesBody = zod.object({
  slackWebhookUrl: zod.string().nullish(),
  standupTone: zod.enum(["professional", "casual", "bullet-points"]).optional(),
  defaultRepo: zod.string().nullish(),
});

export const UpdatePreferencesResponse = zod.object({
  userId: zod.number(),
  slackWebhookUrl: zod.string().nullish(),
  standupTone: zod.enum(["professional", "casual", "bullet-points"]),
  defaultRepo: zod.string().nullish(),
});

/**
 * @summary Test Slack webhook configuration
 */
export const TestSlackWebhookBody = zod.object({
  webhookUrl: zod.string(),
});

export const TestSlackWebhookResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Chat with AI Sprint Coach
 */
export const CoachChatBody = zod.object({
  message: zod.string(),
  conversationHistory: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]),
        content: zod.string(),
      }),
    )
    .optional(),
  repo: zod
    .string()
    .optional()
    .describe("Optional repo to fetch activity context for"),
});

export const CoachChatResponse = zod.object({
  reply: zod.string(),
});

/**
 * @summary Generate a LinkedIn post from a standup
 */
export const GenerateLinkedinPostBody = zod.object({
  standupId: zod.number(),
});

export const GenerateLinkedinPostResponse = zod.object({
  post: zod.string(),
});

/**
 * @summary Get current commit streak
 */
export const GetInsightsStreakResponse = zod.object({
  currentStreak: zod.number(),
  longestStreak: zod.number(),
});

/**
 * @summary Get best day of week by velocity
 */
export const GetInsightsBestDayResponseItem = zod.object({
  day: zod.string(),
  avgVelocity: zod.number(),
  count: zod.number(),
});
export const GetInsightsBestDayResponse = zod.array(
  GetInsightsBestDayResponseItem,
);

/**
 * @summary Get average PR cycle time
 */
export const GetInsightsPrCycleTimeQueryParams = zod.object({
  repo: zod.coerce.string().optional().describe("Full repo name (owner\/repo)"),
});

export const GetInsightsPrCycleTimeResponse = zod.object({
  avgHours: zod.number(),
  label: zod.string(),
  color: zod.string(),
});

/**
 * @summary Get top commit message keywords
 */
export const GetInsightsTopKeywordsResponseItem = zod.object({
  word: zod.string(),
  count: zod.number(),
});
export const GetInsightsTopKeywordsResponse = zod.array(
  GetInsightsTopKeywordsResponseItem,
);